requirements/dev.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile --output-file=- --resolver=backtracking requirements/dev.in
#
anyio==3.6.2
    # via httpcore
build==0.10.0
    # via pip-tools
certifi==2022.12.7
    # via
    #   httpcore
    #   httpx
click==8.1.3
    # via pip-tools
h11==0.14.0
    # via httpcore
httpcore==0.16.3
    # via httpx
httpx==0.23.3
    # via -r requirements/common.in
idna==3.4
    # via
    #   anyio
    #   rfc3986
packaging==23.0
    # via build
pex==2.1.121
    # via -r requirements/dev.in
pillow==9.4.0
    # via -r requirements/common.in
pip-tools==6.12.1
    # via -r requirements/dev.in
pyproject-hooks==1.0.0
    # via build
rfc3986[idna2008]==1.5.0
    # via httpx
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
tomli==2.0.1
    # via
    #   build
    #   pyproject-hooks
wheel==0.38.4
    # via
    #   -r requirements/dev.in
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
